Deprecate `ComposableDependencies`, renamed as `Dependencies`.
