+++ Package.swift
@@ -19,10 +19,6 @@ let package = Package(
       name: "GlobalEnvironment",
       targets: ["GlobalEnvironment"]
     ),
-    .library(
-      name: "GlobalEnvironmentCompat",
-      targets: ["GlobalEnvironmentCompat"]
-    ),
   ],
   dependencies: [
     .package(url: "https://github.com/pointfreeco/swift-composable-architecture", from: "0.21.0"),
@@ -52,12 +48,6 @@ let package = Package(
     ),
     .testTarget(
       name: "GlobalEnvironmentTests",
-      dependencies: [
-        "GlobalEnvironment",
-      ]
-    ),
-    .target(
-      name: "GlobalEnvironmentCompat",
       dependencies: [
         "GlobalEnvironment",
       ]
