+++ Sources/ComposableEnvironment/Deprecations/ComposableDependencies.swift
@@ -1,0 +1,4 @@
+import Dependencies
+
+@available(*, deprecated, renamed: "Dependencies")
+public typealias ComposableDependencies = Dependencies
