+++ Sources/Dependencies/Dependencies.swift
@@ -1,4 +1,19 @@
-public struct DependencyContainer {
+/// This type acts as a namespace to reference your dependencies.
+///
+/// To declare a dependency, create a ``DependencyKey``, and declare a computed property in this
+/// type like you would declare a custom `EnvironmentValue` in SwiftUI. For example, if
+/// `UUIDGeneratorKey` is a ``DependencyKey`` with ``DependencyKey/Value`` == `() -> UUID`:
+/// ```swift
+/// extension Dependencies {
+///   var uuidGenerator: () -> UUID {
+///     get { self[UUIDGeneratorKey.self] }
+///     set { self[UUIDGeneratorKey.self] = newValue }
+///   }
+/// }
+/// ```
+/// This dependency can then be referenced by its keypath `\.uuidGenerator` when invoking the
+/// ``Dependency`` property wrapper.
+public struct Dependencies {
   /// This wrapper enum allows to distinguish dependencies that where defined explicitely for a
   /// given environment from dependencies that were inherited from their parent environment.
   enum DependencyValue {
@@ -34,14 +49,17 @@ public struct DependencyContainer {
     set { values[ObjectIdentifier(key)] = .defined(newValue) }
   }
   
-  public mutating func _mergeFromUpstream(_ upstreamDependencies: DependencyContainer) {
+  /// - Warning: This function is public for implementation reasons. You shouldn't have to call it
+  /// when using the library.
+  public mutating func _mergeFromUpstream(_ upstreamDependencies: Dependencies) {
     // We should preserve dependencies that were defined explicitely.
     for (key, value) in upstreamDependencies.values {
       guard values[key]?.isDefined != true else { continue }
       values[key] = value.inherit()
     }
   }
-  
+  /// - Warning: This function is public for implementation reasons. You shouldn't have to call it
+  /// when using the library.
   public static func _new() -> Self {
     .init()
   }
